Extract tables out of class file
